services/nepali-date-service/pyproject.toml
[project]
name = "nepali-date-service"
version = "0.1.0"
description = "Nepali date conversion microservice"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "nepali-datetime==1.0.8.4",
    "pydantic==2.5.0",
    "python-dotenv==1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "httpx==0.25.2",
]
